QAmergerbranch merging with main.

+++ src/config/firebase.js
@@ -1,8 +1,11 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-import { getAuth, GoogleAuthProvider  } from "firebase/auth";
+import { initializeAppCheck, ReCaptchaV3Provider } from "firebase/app-check"
+import { getAnalytics, logEvent, setUserId, setUserProperties } from "firebase/analytics";
+import { getAuth, GoogleAuthProvider } from "firebase/auth";
+
 import { getFirestore, collection, addDoc, doc, setDoc, serverTimestamp, updateDoc, getDoc, getDocs, deleteDoc, query, where, orderBy, limit, arrayUnion, arrayRemove } from "firebase/firestore";
+
 
 const firebaseConfig = {
     apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
@@ -15,10 +18,22 @@ const firebaseConfig = {
 };
 
 // Initialize Firebase
+
 export const app = initializeApp(firebaseConfig);
+
+const appCheck = initializeAppCheck(app,
+    {
+        provider: new
+            ReCaptchaV3Provider(import.meta.env.VITE_RECAPTCHA_V3_SITE_KEY),
+        isTokenAutoRefreshEnabled: true //Optional, but recommended
+    }
+)
+
 const analytics = getAnalytics(app);
 export const auth = getAuth(app);
 // export const googleProvider = new firebase.auth.GoogleAuthProvider();
 export const googleProvider = new GoogleAuthProvider();
-const db= getFirestore(app);
-export { db, collection, addDoc, doc, setDoc, serverTimestamp, updateDoc, getDoc, getDocs, deleteDoc, query, orderBy, limit, where, arrayUnion, arrayRemove }; 
+
+const db = getFirestore(app);
+export { db, collection, addDoc, doc, setDoc, serverTimestamp, updateDoc, getDoc, getDocs, deleteDoc, query, orderBy, limit, where, arrayUnion, arrayRemove, logEvent, analytics, setUserId, setUserProperties };
+
